+++ example/web_frontend/src/services/interval/index.js
@@ -1,5 +1,28 @@
-module.exports = (compname) => {
+/**
+ * Copyright (c) 2024   Loh Wah Kiang
+ *
+ * openGauss is licensed under Mulan PSL v2.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *
+ *          http://license.coscl.org.cn/MulanPSL2
+ *
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+ * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ * See the Mulan PSL v2 for more details.
+ * -------------------------------------------------------------------------
+ */
+"use strict";
+/**
+ * auth
+ * @module src_services_interval_index
+ */
+module.exports = (...args) => {
   return new Promise(async (resolve, reject) => {
+    const [params, obj] = args;
+    const [pathname, curdir, compname] = params;
+    const [library, sys, cosetting] = obj;
     try {
       let lib = {};
       lib.sleep = (ms, infinate = null) => {
@@ -13,8 +36,8 @@ module.exports = (compname) => {
         c += 1;
       }, 2000);
       resolve(lib);
-    } catch (e) {
-      reject(e);
+    } catch (error) {
+      reject(error);
     }
   });
 };